% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-series.R
\name{caes}
\alias{caes}
\title{Construct aesthetic mappings}
\usage{
caes(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name-value pairs in the form aesthetic = variable.}
}
\value{
a \code{list} of \code{quosure}.
}
\description{
Low-level version of \code{ggplot2::aes}.
}
\examples{
caes(x = month, y = value)
caes(x = month, y = value, fill = city)
}
