% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-demo-data.R
\name{cal-demo-data}
\alias{cal-demo-data}
\alias{cal_demo_data}
\alias{cal_demo_props}
\title{Calendar demo data}
\usage{
cal_demo_data(view = c("month", "week", "day"))

cal_demo_props()
}
\arguments{
\item{view}{Calendar view for which to use the data.}
}
\value{
a \code{data.frame}.
}
\description{
Create calendar demo data for schedules and properties
}
\examples{

# Monthly schedule
cal_demo_data("month")

#' # Weekly schedule
cal_demo_data("week")
}
