% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-demo-data.R
\name{cal_demo_data}
\alias{cal_demo_data}
\title{Create demo schedules data}
\usage{
cal_demo_data(view = c("month", "week", "day"))
}
\arguments{
\item{view}{Calendar view for which to use the data.}
}
\value{
a \code{data.frame}.
}
\description{
Create demo schedules data
}
\examples{

# Monthly schedule
cal_demo_data("month")
}
