% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-proxy.R
\name{cal_proxy_navigation}
\alias{cal_proxy_navigation}
\title{Toggle navigation visibility}
\usage{
cal_proxy_navigation(proxy, navigation)
}
\arguments{
\item{proxy}{A \code{\link[=calendar_proxy]{calendar_proxy()}} \code{htmlwidget} object.}

\item{navigation}{Logical, show or hide the navigation.}
}
\value{
A \code{calendar_proxy} object.
}
\description{
This function can show or hide the calendar navigation.
}
\examples{
library(shiny)
library(toastui)

ui <- fillPage(
  checkboxInput(
    inputId = "cal_navigation",
    label = "Navigation",
    value = TRUE),
  calendarOutput("cal_main")
)

server <- shinyServer(function(input, output) {
  output$cal_main <- renderCalendar({
    calendar(cal_demo_data(), navigation = TRUE)
  })

  observeEvent(input$cal_navigation, {
    cal_proxy_navigation("cal_main", navigation = input$cal_navigation)
  })

})

if (interactive())
  shinyApp(ui, server)
}
\seealso{
Other calendar proxy methods: 
\code{\link{cal_proxy_clear}()},
\code{\link{cal_proxy_options}()},
\code{\link{cal_proxy_toggle}()},
\code{\link{cal_proxy_view}()},
\code{\link{calendar-proxy-navigate}},
\code{\link{calendar-proxy-schedule}},
\code{\link{calendar_proxy}()}
}
\concept{calendar proxy methods}
