% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-proxy.R
\name{cal_proxy_options}
\alias{cal_proxy_options}
\title{Set calendar's options with Proxy}
\usage{
cal_proxy_options(proxy, ...)
}
\arguments{
\item{proxy}{A \code{\link[=calendar_proxy]{calendar_proxy()}} \code{htmlwidget} object.}

\item{...}{Options for the calendar, you can use arguments from \code{\link{calendar}},
\code{\link{cal_month_options}} (under the form \code{month = list(...)}),
\code{\link{cal_week_options}} (under the form \code{week = list(...)})}
}
\value{
No value.
}
\description{
This function allow to set options for a calendar.
}
\seealso{
Other calendar proxy methods: 
\code{\link{cal_proxy_clear}()},
\code{\link{cal_proxy_toggle}()},
\code{\link{cal_proxy_view}()},
\code{\link{calendar-proxy-navigate}},
\code{\link{calendar-proxy-schedule}},
\code{\link{calendar_proxy}()}
}
\concept{calendar proxy methods}
