% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Create an interactive calendar}
\usage{
calendar(
  defaultView = c("month", "week", "day"),
  defaultDate = NULL,
  taskView = FALSE,
  scheduleView = TRUE,
  useDetailPopup = TRUE,
  useCreationPopup = FALSE,
  readOnly = TRUE,
  useNav = FALSE,
  bttnOpts = bttn_options(),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{defaultView}{Default view of calendar. The default value is 'week',
other possible values are 'month' and 'day'.}

\item{defaultDate}{Default date for displaying calendar.}

\item{taskView}{Show the milestone and task in weekly, daily view.
The default value is true. If the value is a vector, it can be 'milestone', 'task'.}

\item{scheduleView}{Show the all day and time grid in weekly, daily view.
The default value is false. If the value is a vector, it can be 'allday', 'time'.}

\item{useDetailPopup}{Logical. Display a pop-up on click with detailled informations about schedules.}

\item{useCreationPopup}{Logical. Allow user to create schedules with a pop-up.}

\item{readOnly}{Calendar is read-only mode and a user can't create and modify any schedule. The default value is true.}

\item{useNav}{Add navigation buttons to got to previous or next period, or return to 'today'.}

\item{bttnOpts}{Options tu customize buttons (only if \code{useNav = TRUE}), see \code{\link{bttn_options}}.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Build interactive calendar with the JavaScript tui-calendar library.
}
