% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calendar_properties}
\alias{calendar_properties}
\title{Calendar properties}
\format{
A \code{data.frame} with 6 rows and 3 variables:
\describe{
\item{Name}{Name of property}
\item{Type}{Type}
\item{Description}{Description}
}
}
\source{
Toast UI documentation (\url{https://nhn.github.io/tui.calendar/latest/CalendarProps/})
}
\usage{
calendar_properties
}
\description{
This dataset contains properties that can be used to set calendars properties in \code{\link{cal_props}}.
}
\keyword{datasets}
