% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-proxy.R
\name{calendar_proxy}
\alias{calendar_proxy}
\title{Proxy for calendar htmlwidget}
\usage{
calendar_proxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
chart to modify (if invoked from a Shiny module, the namespace will be added
automatically).}

\item{session}{the Shiny session object to which the chart belongs; usually the
default value will suffice.}
}
\value{
a `calendar_proxy` object.
}
\description{
Proxy for calendar htmlwidget
}
