% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart-options.R
\name{chart_options}
\alias{chart_options}
\title{Chart options}
\usage{
chart_options(.chart, ...)
}
\arguments{
\item{.chart}{A \code{chart} htmlwidget.}

\item{...}{Named list of options, options depends on chart's type,
see common options \href{https://github.com/nhn/tui.chart/blob/main/docs/en/common-chart-options.md}{here}.}
}
\value{
A \code{chart} htmlwidget.
}
\description{
Chart options
}
\examples{
chart(mtcars, caes(x = mpg, y = wt), type = "scatter") \%>\% 
  chart_options(
    chart = list(title = "A scatter chart")
  )
}
