% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{datagrid}
\alias{datagrid}
\title{Interactive tables with tui-grid}
\usage{
datagrid(
  data,
  ...,
  sortable = TRUE,
  pagination = NULL,
  filters = FALSE,
  cols_width = "auto",
  theme = c("clean", "striped", "default"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} or something convertible in \code{data.frame}.}

\item{...}{Arguments passed to the \code{Grid} JavaScript method : \url{https://nhn.github.io/tui.grid/latest/Grid}.}

\item{sortable}{Logical, allow to sort columns.}

\item{pagination}{Number of rows per page to display, default to \code{NULL} (no pagination).}

\item{filters}{Logical, allow to filter columns.}

\item{cols_width}{Width for the columns, can be \code{"auto"} (width is determined by column's content)
or a single or numeric vector to set the width in pixel. Use \code{NULL} to disable and have equal width columns.}

\item{theme}{Set styles for the entire table.}

\item{width, height}{Width and height of the table in a CSS unit or a numeric.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Create interactive tables : sortable, filterable,
 editable with the JavaScript library \href{https://ui.toast.com/tui-grid/}{tui-grid}.
}
\examples{
library(toastui)

# default usage
datagrid(rolling_stones_50)

# disable sorting
datagrid(rolling_stones_50, sortable = FALSE)

# enable default filtering
datagrid(rolling_stones_50, filters = TRUE)

# enable pagination (10 rows per page)
datagrid(rolling_stones_50, pagination = 10)

# Themes
datagrid(rolling_stones_50, theme = "striped")
datagrid(rolling_stones_50, theme = "default")


# Empty table
datagrid(list())

# with columns
datagrid(data.frame(
  variable_1 = character(0),
  variable_2 = character(0)
))



}
