% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-style.R
\name{grid-cell-style}
\alias{grid-cell-style}
\alias{grid_cell_style}
\alias{grid_cells_style}
\title{Set grid cell(s) style}
\usage{
grid_cell_style(
  grid,
  expr,
  column,
  background = NULL,
  color = NULL,
  fontWeight = NULL,
  ...,
  class = NULL,
  cssProperties = NULL
)

grid_cells_style(
  grid,
  fun,
  columns,
  background = NULL,
  color = NULL,
  ...,
  class = NULL,
  cssProperties = NULL
)
}
\arguments{
\item{grid}{A grid created with \code{\link{datagrid}}.}

\item{expr}{An expression giving position of row. Must return a logical vector.}

\item{column}{Name of column (variable name) to identify cells to style.}

\item{background}{Background color.}

\item{color}{Text color.}

\item{fontWeight}{Font weight, you can use \code{"bold"} for example.}

\item{...}{Other CSS properties.}

\item{class}{CSS class to apply to the row.}

\item{cssProperties}{Alternative to specify CSS properties with a named list.}

\item{fun}{Function to apply to \code{columns} to identify rows to style.}

\item{columns}{Columns names to use with \code{fun}.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Customize cell(s) appearance with CSS
 according to an expression in the data used in the grid.
}
\examples{
library(toastui)

datagrid(iris) \%>\%
  grid_cell_style(
    Sepal.Length > 5,
    column = "Sepal.Length",
    background = "#F781BE",
    fontWeight = "bold"
  )


dat <- iris[c(1:3, 51:53, 101:103), ]
datagrid(dat) \%>\%
  grid_cell_style(
    Species == "setosa",
    column = "Species",
    background = "#E41A1C80",
    color = "#FFF"
  ) \%>\%
  grid_cell_style(
    Species == "versicolor",
    column = "Species",
    background = "#377EB880"
  ) \%>\%
  grid_cell_style(
    Species == "virginica",
    column = "Species",
    background = "#4DAF4A80"
  )


# Use rlang to use character
library(rlang)
my_var <- "Sepal.Length"
datagrid(iris) \%>\%
  grid_cell_style(
    !!sym(my_var) > 5,
    column = "Sepal.Length",
    background = "#F781BE"
  )




# Style multiple columns

cor_longley <- as.data.frame(cor(longley))
cor_longley$Var <- row.names(cor_longley)
vars <- c("GNP.deflator", "GNP",
          "Unemployed", "Armed.Forces",
          "Population", "Year", "Employed")
datagrid(cor_longley[, c("Var", vars)]) \%>\%
  grid_cells_style(
    fun = ~ . > 0.9,
    columns = vars,
    background = "#053061",
    color = "#FFF"
  ) \%>\%
  grid_cells_style(
    fun = ~ . > 0 & . <= 0.9,
    columns = vars,
    background = "#539dc8",
    color = "#FFF"
  ) \%>\%
  grid_cells_style(
    fun = ~ . < 0,
    columns = vars,
    background = "#b51f2e",
    color = "#FFF"
  )












}
