% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-style.R
\name{grid_cell_style}
\alias{grid_cell_style}
\title{Set grid cell style}
\usage{
grid_cell_style(
  grid,
  expr,
  column,
  background = NULL,
  color = NULL,
  ...,
  class = NULL
)
}
\arguments{
\item{grid}{A grid created with \code{\link{tuigrid}}.}

\item{expr}{An expression giving position of row. Must return a logical vector.}

\item{column}{Name of column (variable name) to identify cells to style.}

\item{background}{Background color.}

\item{color}{Text color.}

\item{...}{Other CSS properties.}

\item{class}{CSS class to apply to the row.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Set grid cell style
}
\examples{
library(tuigridr)

tuigrid(iris) \%>\%
  grid_cell_style(
    Sepal.Length > 5,
    column = "Sepal.Length",
    background = "#F781BE"
  )


dat <- iris[c(1:3, 51:53, 101:103), ]
tuigrid(dat) \%>\%
  grid_cell_style(
    Species == "setosa",
    column = "Species",
    background = "#E41A1C80",
    color = "#FFF"
  ) \%>\%
  grid_cell_style(
    Species == "versicolor",
    column = "Species",
    background = "#377EB880"
  ) \%>\%
  grid_cell_style(
    Species == "virginica",
    column = "Species",
    background = "#4DAF4A80"
  )


# Use rlang to use character
library(rlang)
my_var <- "Sepal.Length"
tuigrid(iris) \%>\%
  grid_cell_style(
    !!sym(my_var) > 5,
    column = "Sepal.Length",
    background = "#F781BE"
  )
}
