% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-editor.R
\name{grid_editor}
\alias{grid_editor}
\title{Grid editor for columns}
\usage{
grid_editor(
  grid,
  column,
  type = c("text", "checkbox", "select", "radio", "password"),
  choices = NULL
)
}
\arguments{
\item{grid}{A table created with \code{\link{datagrid}}.}

\item{column}{Column for which to activate the editable content.}

\item{type}{Type of editor: \code{"text"}, \code{"checkbox"},
\code{"select"}, \code{"radio"} or \code{"password"}.}

\item{choices}{Vector of choices, required for \code{"checkbox"},
\code{"select"} and \code{"radio"} type.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Allow to edit content of columns with different inputs,
 then retrieve value server-side in shiny application with \code{input$<outputId>_data}.
}
\examples{
library(toastui)

# Example data
dat <- data.frame(
  character = month.name,
  select = month.name,
  checkbox = month.abb,
  radio = month.name,
  password = month.name
)


datagrid(dat) \%>\% 
  grid_editor(
    column = "character",
    type = "text"
  ) \%>\% 
  grid_editor(
    column = "select",
    type = "select",
    choices = month.name
  ) \%>\% 
  grid_editor(
    column = "checkbox",
    type = "checkbox",
    choices = month.abb
  ) \%>\% 
  grid_editor(
    column = "radio",
    type = "radio",
    choices = month.name
  ) \%>\% 
  grid_editor(
    column = "password",
    type = "password"
  )

}
\seealso{
\code{\link{grid_editor_date}} for a date picker.
}
