% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-editor.R
\name{grid_editor_opts}
\alias{grid_editor_opts}
\title{Grid editor options}
\usage{
grid_editor_opts(
  grid,
  editingEvent = c("dblclick", "click"),
  updateOnClick = NULL
)
}
\arguments{
\item{grid}{A table created with \code{\link{datagrid}}.}

\item{editingEvent}{If set to \code{"click"}, editable cell in
the view-mode will be changed to edit-mode by a single click.}

\item{updateOnClick}{Use an \code{actionButton} inputId to send
edited data to the server only on click.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Grid editor options
}
