% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{grid_header}
\alias{grid_header}
\title{Header options}
\usage{
grid_header(
  grid,
  complexColumns = NULL,
  align = NULL,
  valign = NULL,
  height = NULL
)
}
\arguments{
\item{grid}{A table created with \code{\link{tuigrid}}.}

\item{complexColumns}{\code{list}. This options creates new parent
headers of the multiple columns which includes the headers of
specified columns, and sets up the hierarchy.}

\item{align}{Horizontal alignment of the header content.
Available values are 'left', 'center', 'right'.}

\item{valign}{Vertical alignment of the row header content.
Available values are 'top', 'middle', 'bottom'.}

\item{height}{Numeric. The height of the header area.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Header options
}
\examples{

tuigrid(iris) \%>\%
  grid_header(
    align = "left",
    height = "150px"
  )


tuigrid(iris, bodyHeight = "auto", height = 500) \%>\%
  grid_header(
    complexColumns = list(
      list(
        header = "Sepal",
        name = "Sepal",
        childNames = c("Sepal.Length", "Sepal.Width")
      ),
      list(
        header = "Petal",
        name = "Petal",
        childNames = c("Petal.Length", "Petal.Width")
      )
    ),
    height = 80,
    valign = "middle"
  )

}
