% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-proxy.R
\name{grid_proxy_addrow}
\alias{grid_proxy_addrow}
\title{Add rows to a existent \code{datagrid}}
\usage{
grid_proxy_addrow(proxy, data)
}
\arguments{
\item{proxy}{A \code{\link{grid_proxy}} or \code{outputId} of the grid.}

\item{data}{\code{data.frame} to append in the grid.}
}
\value{
No value.
}
\description{
Add rows to a existent \code{datagrid}
}
\examples{
library(shiny)

ui <- fluidPage(
  tags$h2("Append row to grid"),
  datagridOutput("grid"),
  actionButton(
    inputId = "add", 
    label = "Add row", 
    class = "btn-block"
  )
)

server <- function(input, output, session) {
  
  dat <- data.frame(
    character = month.name,
    select = month.name,
    checkbox = month.abb,
    radio = month.name,
    password = month.name
  )
  
  output$grid <- renderDatagrid({
    datagrid(rolling_stones_50[1, ])
  })
  
  value <- reactiveVal(1)
  observeEvent(input$add, {
    row <- value() + 1
    grid_proxy_addrow(
      proxy = "grid",
      data = rolling_stones_50[row, ]
    )
    value(row)
  })
  
}

if (interactive())
  shinyApp(ui, server)
}
