% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-attributes.R
\name{grid_row_merge}
\alias{grid_row_merge}
\title{Merge rows}
\usage{
grid_row_merge(grid, vars)
}
\arguments{
\item{grid}{A grid created with \code{\link{tuigrid}}.}

\item{vars}{Variable(s) in which merge consecutives rows.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Merge rows
}
\examples{
library(tuigridr)

dat <- iris[c(1:3, 51:53, 101:103), ]

tuigrid(dat) \%>\%
  grid_row_merge(vars = "Species")

tuigrid(dat) \%>\%
  grid_row_merge(vars = "Species") \%>\%
  grid_row_merge(vars = "Petal.Width")



}
