% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-style.R
\name{grid_row_style}
\alias{grid_row_style}
\title{Set grid row style}
\usage{
grid_row_style(grid, expr, background = NULL, color = NULL, ..., class = NULL)
}
\arguments{
\item{grid}{A grid created with \code{\link{datagrid}}.}

\item{expr}{An expression giving position of row. Must return a logical vector.}

\item{background}{Background color.}

\item{color}{Text color.}

\item{...}{Other CSS properties.}

\item{class}{CSS class to apply to the row.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Set grid row style
}
\examples{
library(toastui)

datagrid(iris) \%>\%
  grid_row_style(
    Sepal.Length > 5,
    background = "#F781BE"
  )


dat <- iris[c(1:3, 51:53, 101:103), ]
datagrid(dat) \%>\%
  grid_row_style(
    Species == "setosa",
    background = "#E41A1C80",
    color = "#FFF"
  ) \%>\%
  grid_row_style(
    Species == "versicolor",
    background = "#377EB880"
  ) \%>\%
  grid_row_style(
    Species == "virginica",
    background = "#4DAF4A80"
  )


# Use rlang to use character
library(rlang)
my_var <- "Sepal.Length"
datagrid(iris) \%>\%
  grid_row_style(
    !!sym(my_var) > 5,
    background = "#F781BE"
  )




}
