% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-sparkline.R
\name{grid_sparkline}
\alias{grid_sparkline}
\title{Title}
\usage{
grid_sparkline(grid, column, renderer, height = "40px", styles = NULL)
}
\arguments{
\item{renderer}{}
}
\value{

}
\description{
Title
}
