% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{grid_theme}
\alias{grid_theme}
\title{Modify theme options}
\usage{
grid_theme(
  grid,
  row = NULL,
  cell = NULL,
  area = NULL,
  outline = NULL,
  selection = NULL,
  scrollbar = NULL,
  frozenBorder = NULL,
  ...
)
}
\arguments{
\item{grid}{A grid created with \code{\link{tuigrid}}.}

\item{row}{Styles for the table rows.}

\item{cell}{Styles for the table cells.}

\item{area}{Styles for the table areas.}

\item{outline}{Styles for the table outline.}

\item{selection}{Styles for a selection layer.}

\item{scrollbar}{Styles for scrollbars.}

\item{frozenBorder}{Styles for a frozen border.}

\item{...}{Additional arguments.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Properties to customize grid theme, each argument
 is a list of parameters, see full list here :
 \url{https://nhn.github.io/tui.grid/latest/Grid#applyTheme}.
}
\examples{

# Default
tuigrid(iris)

# Customize theme
tuigrid(iris) \%>\%
  grid_theme(
    row = list(
      even = list(
        background = "#ddebf7"
      )
    ),
    cell = list(
      normal = list(
        border = "#9bc2e6",
        showHorizontalBorder = TRUE
      ),
      header = list(
        background = "#5b9bd5",
        text = "#FFF"
      )
    )
  )




}
