% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-validation.R
\name{grid_validation}
\alias{grid_validation}
\title{Grid validation options for columns}
\usage{
grid_validation(
  grid,
  column,
  required = NULL,
  type = NULL,
  min = NULL,
  max = NULL,
  regExp = NULL,
  jsfun = NULL
)
}
\arguments{
\item{grid}{A table created with \code{\link{datagrid}}.}

\item{column}{Column for which to activate the content validation.}

\item{required}{If set to \code{TRUE}, the data of the column will be checked to be not empty.}

\item{type}{Type of data, can be \code{"string"} or \code{"number"}.}

\item{min}{For numeric values, the minimum acceptable value.}

\item{max}{For numeric values, the maximum acceptable value.}

\item{regExp}{A regular expression to validate content.}

\item{jsfun}{A \code{JS} function to validate content.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Validate columns' content with rules, useful when content is editable.
}
\examples{
library(toastui)

dat <- data.frame(
  number = c("a", 2:10),
  character = c("A", "B", "C", character(7)),
  "min=0 & max=100" = sample(1:150, 10),
  email = c("victor@mail.com", "victor"),
  check.names = FALSE
)


datagrid(dat) \%>\% 
  grid_editor(
    column = "number",
    type = "text"
  ) \%>\% 
  grid_validation(
    column = "number", 
    type = "number"
  ) \%>\% 
  grid_editor(
    column = "character",
    type = "text"
  ) \%>\% 
  grid_validation(
    column = "character", 
    type = "string", 
    required = TRUE
  ) \%>\% 
  grid_editor(
    column = "min=0 & max=100",
    type = "text"
  ) \%>\% 
  grid_validation(
    column = "min=0 & max=100", 
    type = "number",
    min = 0, max = 100
  ) \%>\%
  grid_editor(
    column = "email",
    type = "text"
  ) \%>\%
  grid_validation(
    column = "email",
    type = "string",
    regExp = "^([a-zA-Z0-9_\\\\-\\\\.]+)@([a-zA-Z0-9_\\\\-\\\\.]+)\\\\.([a-zA-Z]{2,5})$"
  )

}
