% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrid.R
\name{guess_colwidths_options}
\alias{guess_colwidths_options}
\title{Options for guessing columns widths}
\usage{
guess_colwidths_options(min_width = 70, max_width = 400, mul = 1, add = 0)
}
\arguments{
\item{min_width}{Minimal width.}

\item{max_width}{Maximal width.}

\item{mul}{Multiplicative constant.}

\item{add}{Additive constant}
}
\value{
a \code{list} of options to use in \code{\link[=datagrid]{datagrid()}}.
}
\description{
Options for guessing columns widths
}
\examples{
datagrid(rolling_stones_50, colwidths = "guess")
datagrid(
  rolling_stones_50,
  colwidths = "guess",
  guess_colwidths_opts= guess_colwidths_options(mul = 2)
)
}
