% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{met_paris}
\alias{met_paris}
\title{Meteorological for Le Bourget Station}
\format{
A \code{data.frame} with 12 rows and 3 variables:
\describe{
\item{month}{Month of the year}
\item{temp}{List column containing data.frame with 2 column "date  and"temp"}
\item{rh}{List column containing data.frame with 2 column "date  and"rh"}
}
}
\source{
Data collected with package stationaRy from NOAA
}
\usage{
met_paris
}
\description{
This dataset contains temperature and relative humidity for year 2020.
}
\keyword{datasets}
