% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar-proxy.R
\name{proxy-schedule}
\alias{proxy-schedule}
\alias{cal_proxy_add}
\alias{cal_proxy_delete}
\alias{cal_proxy_update}
\title{Create / Update / Delete schedule(s) with Proxy}
\usage{
cal_proxy_add(proxy, value)

cal_proxy_delete(proxy, value)

cal_proxy_update(proxy, value)
}
\arguments{
\item{proxy}{A \code{\link{calendar_proxy}} \code{htmlwidget} object.}

\item{start}{The start time.}

\item{end}{The end time.}

\item{title}{The schedule title.}

\item{body}{The schedule body text which is text/plain.}

\item{id}{An id for the schedule.}

\item{calendarId}{An id for the calendar.}

\item{category}{The schedule type ('milestone', 'task', allday', 'time').}

\item{...}{Additionnal arguments passed to the JavaScript method.}

\item{.list}{A \code{list} with same information as above, useful
with \code{input$<outputId>_<add/update/delete>_schedule}.}
}
\value{
No value.
}
\description{
These functions allow to create new schedule(s), update existing
 ones and delete schedule in a calendar within the server in a Shiny application.
}
