% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-proxy.R
\name{proxy_grid_addrow}
\alias{proxy_grid_addrow}
\title{Add rows to a existant \code{datagrid}}
\usage{
proxy_grid_addrow(proxy, data)
}
\arguments{
\item{proxy}{A \code{\link{grid_proxy}} or \code{outputId} of the grid.}

\item{data}{\code{data.frame} to append in the grid.}
}
\value{
No value.
}
\description{
Add rows to a existant \code{datagrid}
}
