% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rolling_stones_500}
\alias{rolling_stones_500}
\title{Rolling Stone's 500 Greatest Albums of All Time}
\format{
A \code{data.frame} with 500 rows and 6 variables:
\describe{
\item{Number}{Position on the list}
\item{Year}{Year of release}
\item{Album}{Album name}
\item{Artist}{Artist name}
\item{Genre}{Genre name}
\item{Subgenre}{Subgenre name}
}
}
\source{
Gibs on Kaggle (https://www.kaggle.com/datasets/notgibs/500-greatest-albums-of-all-time-rolling-stone/)
}
\usage{
rolling_stones_500
}
\description{
Data about Rolling Stone magazine's (2012) top 500 albums of all time list.
}
\keyword{datasets}
