% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedules_properties}
\alias{schedules_properties}
\title{Schedules properties}
\format{
A \code{data.frame} with 26 rows and 3 variables:
\describe{
\item{Name}{Name of property}
\item{Type}{Type}
\item{Description}{Description}
}
}
\source{
Toast UI documentation (\url{https://nhn.github.io/tui.calendar/latest/EventObject/})
}
\usage{
schedules_properties
}
\description{
This dataset contains properties that can be use to create schedules in \code{\link[=calendar]{calendar()}}.
}
\keyword{datasets}
