% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_row_style.R
\name{set_row_style}
\alias{set_row_style}
\title{Set row style}
\usage{
set_row_style(grid, expr, background = NULL, color = NULL, ..., class = NULL)
}
\arguments{
\item{grid}{A grid created with \code{\link{tuigrid}}.}

\item{expr}{An expression giving positionof row}

\item{background}{Background color.}

\item{color}{Text color.}

\item{...}{CSS properties.}

\item{class}{CSS class to apply to the row.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Set row style
}
\examples{

tuigrid(iris) \%>\%
  set_row_style(
    Sepal.Length > 5,
    background = "#F781BE"
  )


dat <- iris[c(1:3, 51:53, 101:103), ]
tuigrid(dat) \%>\%
  set_row_style(
    Species == "setosa",
    background = "#E41A1C80",
    color = "#FFF"
  ) \%>\%
  set_row_style(
    Species == "versicolor",
    background = "#377EB880"
  ) \%>\%
  set_row_style(
    Species == "virginica",
    background = "#4DAF4A80"
  )
}
