% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastui-package.R
\name{toastui-exports}
\alias{toastui-exports}
\alias{\%>\%}
\alias{JS}
\title{toastui exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the toastui package to avoid listing the magrittr
as Depends of toastui
}
