% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastui-package.R
\docType{package}
\name{toastui}
\alias{toastui}
\title{HTMLwidget interface to the
\href{https://ui.toast.com/}{TOASTUI} javascript libraries.}
\description{
Create interactive tables, calendars and charts with one package.
}
\section{Tables}{

Interactive and editable tables with \href{https://ui.toast.com/tui-grid/}{tui-grid}, see \code{\link{datagrid}}.
}

\section{Calendars}{

Interactive and editable calendars with \href{https://ui.toast.com/tui-calendar/}{tui-calendar}, see .
}

\section{Charts}{

Interactive charts with \href{https://ui.toast.com/tui-chart/}{tui-chart}, see .
}

\author{
Victor Perrier (@dreamRs_fr)
}
