% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuigridr-package.R
\name{tuigridr-exports}
\alias{tuigridr-exports}
\alias{\%>\%}
\title{tuigridr exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the tuigridr package to avoid listing the magrittr
as Depends of tuigridr.
}
