% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuigridr-package.R
\docType{package}
\name{tuigridr}
\alias{tuigridr}
\title{An \code{htmlwidget} interface to the
tui-grid javascript library}
\description{
This package allow you to use tui-grid.js (\url{https://ui.toast.com/tui-grid/}),
to create interactive tables.
}
\author{
Victor Perrier (@dreamRs_fr)
}
