% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{bttn_options}
\alias{bttn_options}
\title{Options for buttons displayed above calendar}
\usage{
bttn_options(
  today_label = "Today",
  prev_label = tags$i(class = "fa fa-chevron-left"),
  next_label = tags$i(class = "fa fa-chevron-right"),
  class = "bttn-jelly bttn-sm bttn-primary",
  bg = NULL,
  color = NULL
)
}
\arguments{
\item{today_label}{Text to display on today button.}

\item{prev_label}{Text to display on prev button.}

\item{next_label}{Text to display on next button.}

\item{class}{Class to add to buttons.}

\item{bg}{Background color.}

\item{color}{Text color.}
}
\value{
a \code{list}.
}
\description{
Options for buttons displayed above calendar
}
\note{
Buttons are generated with the following CSS library : \url{http://bttn.surge.sh/},
 where you can find available options for \code{class} argument.
}
\examples{
# Use another button style
calendar(
  defaultView = "month", useNav = TRUE,
  bttnOpts = bttn_options(
    class = "bttn-stretch bttn-sm bttn-warning"
  )
)

# Custom colors (background and text)
calendar(
  defaultView = "month", useNav = TRUE,
  bttnOpts = bttn_options(bg = "#FE2E2E", color = "#FFF")
)

# both
calendar(
  defaultView = "month", useNav = TRUE,
  bttnOpts = bttn_options(
    bg = "#04B431", color = "#FFF",
    class = "bttn-float bttn-md"
  )
)
}
