% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{grid_filters}
\alias{grid_filters}
\title{Set filters options}
\usage{
grid_filters(
  grid,
  vars,
  showApplyBtn = NULL,
  showClearBtn = NULL,
  operator = NULL,
  format = "yyyy-MM-dd",
  type = c("auto", "text", "date", "number", "select")
)
}
\arguments{
\item{grid}{A table created with \code{\link{tuigrid}}.}

\item{vars}{Name(s) of column in the data used in \code{\link{tuigrid}}.}

\item{showApplyBtn}{Apply filters only when button is pressed.}

\item{showClearBtn}{Reset the filter that has already been applied.}

\item{operator}{Multi-option filter, the operator used against multiple rules : \code{"OR"} or \code{"AND"}.}

\item{format}{Date format.}

\item{type}{Type of filter : \code{"auto"}, \code{"text"}, \code{"number"}, \code{"date"} or \code{"select"}.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Set filters options
}
