% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-row-attributes.R
\name{grid_row_merge}
\alias{grid_row_merge}
\title{Merge rows}
\usage{
grid_row_merge(grid, vars)
}
\arguments{
\item{grid}{A grid created with \code{\link{datagrid}}.}

\item{vars}{Variable(s) in which merge consecutives rows.}
}
\value{
A \code{datagrid} htmlwidget.
}
\description{
Merge rows
}
\examples{
library(toastui)

dat <- iris[c(1:3, 51:53, 101:103), ]

datagrid(dat) \%>\%
  grid_row_merge(vars = "Species")

datagrid(dat) \%>\%
  grid_row_merge(vars = "Species") \%>\%
  grid_row_merge(vars = "Petal.Width")



}
