% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{grid_row_selection}
\alias{grid_row_selection}
\title{Row selection (in shiny)}
\usage{
grid_row_selection(
  grid,
  inputId,
  label = NULL,
  return = c("data", "index"),
  width = NULL
)
}
\arguments{
\item{grid}{A table created with \code{\link{tuigrid}}.}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label in header row.}

\item{return}{Value that will be accessible via \code{input} :
a \code{data.frame} with selected row(s) or just the index of rows.}

\item{width}{Width of the column.}
}
\value{
A \code{tuidgridr} htmlwidget.
}
\description{
Row selection (in shiny)
}
\examples{
if (interactive()) {
  library(shiny)
  library(tuigridr)

  ui <- fluidPage(
    tags$h2("tuigridr row selection"),
    tuigridOutput("grid"),
    verbatimTextOutput("res")
  )

  server <- function(input, output, session) {

    df <- data.frame(
      index = 1:12,
      month = month.name,
      letters = letters[1:12]
    )

    output$grid <- renderTuigrid({
      tuigrid(df) \%>\%
        grid_row_selection(
          inputId = "rows", width = 50
        )
    })

    output$res <- renderPrint({
      input$rows
    })
  }

  shinyApp(ui, server)
}
}
