% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-language.R
\name{set_grid_lang}
\alias{set_grid_lang}
\title{Set grid language options}
\usage{
set_grid_lang(
  display.noData = "No data",
  display.loadingData = "Loading data...",
  display.resizeHandleGuide = NULL
)
}
\arguments{
\item{display.noData, display.loadingData, display.resizeHandleGuide}{Language options.}
}
\value{
No return value.
}
\description{
Set grid language options
}
\examples{
library(toastui)
set_grid_lang(display.noData = "Pas de donn\u00e9es")

datagrid(data.frame(
  variable_1 = character(0),
  variable_2 = character(0)
))

}
