% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastui-package.R
\docType{package}
\name{toastui}
\alias{toastui-package}
\alias{toastui}
\title{HTMLwidget interface to the
\href{https://ui.toast.com/}{TOASTUI} javascript libraries.}
\description{
Create interactive tables, calendars and charts with one package.
}
\section{Tables}{

Interactive and editable tables with \href{https://ui.toast.com/tui-grid/}{tui-grid}, see \code{\link[=datagrid]{datagrid()}}.
}

\section{Calendars}{

Interactive and editable calendars with \href{https://ui.toast.com/tui-calendar/}{tui-calendar}, see \code{\link{calendar}}.
}

\section{Charts}{

Interactive charts with \href{https://ui.toast.com/tui-chart/}{tui-chart}, see \code{\link{chart}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://dreamrs.github.io/toastui/}
  \item Report bugs at \url{https://github.com/dreamRs/toastui/issues}
}

}
\author{
Victor Perrier (@dreamRs_fr)
}
