% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuigridr.R
\name{tuigrid}
\alias{tuigrid}
\title{Interactive table with tui-grid}
\usage{
tuigrid(
  data,
  theme = c("clean", "striped", "default"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} or something convertible en \code{data.frame}.}

\item{theme}{Set styles for the entire table.}
}
\description{
Create interactive table : sortable, filterable,
 editable with the JavaScript library \href{https://ui.toast.com/tui-grid/}{tui-grid}.
}
