% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuigridr.R
\name{tuigrid}
\alias{tuigrid}
\title{Interactive table with tui-grid}
\usage{
tuigrid(
  data,
  ...,
  sortable = TRUE,
  pagination = NULL,
  filters = FALSE,
  theme = c("clean", "striped", "default"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} or something convertible en \code{data.frame}.}

\item{...}{Arguments passed to the \code{Grid} JavaScript method : \url{https://nhn.github.io/tui.grid/latest/Grid}.}

\item{sortable}{Logical, allow to sort columns.}

\item{pagination}{Number of rows per page to display, default to \code{NULL} (no pagination).}

\item{filters}{Logical, allow to filter columns.}

\item{theme}{Set styles for the entire table.}

\item{width, height}{Width and height of the table in a CSS unit or a numeric.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Create interactive table : sortable, filterable,
 editable with the JavaScript library \href{https://ui.toast.com/tui-grid/}{tui-grid}.
}
